Algumas mudanças no que ja tinha

+++ Assets/_Game/Scripts/Cutscenes/Cutscene.cs
@@ -50,7 +50,7 @@ public class Cutscene : MonoBehaviour
         DeleteCharactersWithName();
     }
 
-    public void PlayCutsceneIfTriggered(int cutsceneNum)
+    public void PlayCutsceneIfTriggered(float cutsceneNum)
     {
         if (cutsceneNum == this.cutsceneNum)
             playableDirector.Play();
